refactor: isolate command palette runtime
