+++ layouts/_partials/search-input.html
@@ -27,7 +27,7 @@
 {{- end -}}
 
 
-{{ if .Site.Params.offlineSearch -}}
+{{ if partial "shell/search-enabled.html" . -}}
 {{ .Scratch.Add "oink-search" 1 -}}
 
 {{ $offlineSearchLink := partialCached "offline-search-index.html" . .Site.Language.Lang -}}
@@ -48,7 +48,7 @@
       It causes the json file loading error when when relativeURLs is enabled.
       https://github.com/google/docsy/issues/181
     */}}
-    data-offline-search-index-json-src="{{ $offlineSearchLink }}"
+    data-td-shell-search-open
     data-offline-search-base-href="/"
     data-offline-search-max-results="{{ .Site.Params.offlineSearchMaxResults | default 10 }}"
   >
