+++ scripts/check-prd4-runtime.py
@@ -1,0 +1,106 @@
+#!/usr/bin/env python3
+"""Check PRD 4 runtime isolation and transient-surface wiring."""
+
+from __future__ import annotations
+
+import re
+import sys
+from pathlib import Path
+
+
+ROOT = Path(__file__).resolve().parents[1]
+
+
+def read(path: str) -> str:
+    return (ROOT / path).read_text(encoding="utf-8")
+
+
+def require(condition: bool, message: str) -> None:
+    if not condition:
+        raise RuntimeError(message)
+
+
+def main() -> int:
+    try:
+        search_enabled = read("layouts/_partials/shell/search-enabled.html")
+        head = read("layouts/_partials/head.html")
+        scripts = read("layouts/_partials/scripts.html")
+        search_input = read("layouts/_partials/search-input.html")
+        docs_shell = read("assets/js/docs-shell.js")
+        palette = read("assets/js/command-palette.js")
+        coordinator = read("assets/js/surface-coordinator.js")
+        base = read("assets/js/base.js")
+
+        require(
+            '.Site.Params.offlineSearch' in search_enabled
+            and 'partial "shell/chrome-enabled.html"' in search_enabled
+            and 'partial "outputformat.html"' in search_enabled
+            and '(ne $format "print")' in search_enabled,
+            "search-enabled.html is not the complete capability predicate",
+        )
+        require(
+            '{{ if partial "shell/search-enabled.html" . -}}' in head,
+            "Lunr is not gated by the canonical capability",
+        )
+        require(
+            '{{ if $localSearch }}{{ $jsArray = $jsArray | append $jsPalette }}'
+            in scripts,
+            "Palette controller is not gated by the canonical capability",
+        )
+        require(
+            'resources.Get "js/offline-search.js"' not in scripts,
+            "legacy offline-search runtime remains in the shared bundle",
+        )
+        require(
+            'partial "shell/search-enabled.html"' in search_input
+            and "data-td-shell-search-open" in search_input,
+            "legacy search-input partial does not delegate to the Palette",
+        )
+        require(
+            "$hasGoogleSearch" in scripts and ".Site.Language.Lang" in scripts,
+            "bundle key is missing search or language discriminators",
+        )
+        require(
+            "function initSearch()" not in docs_shell
+            and "td-shell-search" not in docs_shell,
+            "search controller remains coupled to docs-shell.js",
+        )
+        require(
+            "function initSearch()" in palette
+            and "CJK =" in palette
+            and "metaKey || e.ctrlKey" in palette
+            and "tabTrap(panel, isOpen)" in palette,
+            "extracted Palette lost search or keyboard behavior",
+        )
+        require(
+            all(
+                token in coordinator
+                for token in ("register", "closeOthers", "closeAll")
+            ),
+            "surface coordinator API is incomplete",
+        )
+        for surface in (
+            "palette",
+            "drawer",
+            "root-menu",
+            "mobile-menu",
+            "language-menu-",
+            "version-menu-",
+        ):
+            require(
+                surface in palette + docs_shell + base,
+                f"surface coordinator is missing {surface}",
+            )
+        require(
+            not re.search(r"\x1b|\r", palette + coordinator),
+            "runtime source contains terminal/control characters",
+        )
+    except (OSError, RuntimeError) as exc:
+        print(f"PRD 4 runtime check failed: {exc}", file=sys.stderr)
+        return 1
+    print("PRD 4 runtime isolation checks passed")
+    return 0
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
